backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
httpx
google-genai
python-dotenv
sentence-transformers
sqlalchemy
pymysql
passlib[bcrypt]
python-jose[cryptography]
email-validator
